% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{are_anagrams}
\alias{are_anagrams}
\title{returns TRUE if both strings are anagrams and not the same}
\usage{
are_anagrams(w1, w2)
}
\arguments{
\item{w1}{string 1}

\item{w2}{string 2}
}
\value{
TRUE or FALSE
}
\description{
returns TRUE if both strings are anagrams and not the same
}
\author{
Walter van Heuven
}
