% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{bigrams}
\alias{bigrams}
\title{bigrams}
\usage{
bigrams(the_str, type = "adjacent", max_distance = -1)
}
\arguments{
\item{the_str}{letter string}

\item{type}{'all', 'adjacent', 'open'}

\item{max_distance}{distance between the letters of the open bigrams. -1 no restriction for the number of open bigrams}
}
\value{
list of characters
}
\description{
`bigrams` returns the bigrams of a letter string
}
\details{
strngrams

Written by Walter van heuven, https://waltervanheuven.net

Either all possible bigrams or only adjacent bigrams or open bigrams are returned
}
\examples{
\dontrun{
str_bigram("DREAM")
str_bigram("DREAM", "all")
}
}
\author{
Walter van Heuven
}
