% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{gtzero}
\alias{gtzero}
\title{gtzero}
\usage{
gtzero(the_str, bigram_table)
}
\arguments{
\item{the_str}{letter string}

\item{bigram_table}{bigram table}
}
\value{
number of bigrams
}
\description{
`gtzero` returns the number of bigrams in letter string that exist in the bigram_table.
}
\details{
Based on definition of Mendelsohn (1976),
}
\references{
\url{https://link.springer.com/article/10.3758/BF03213228}
}
\author{
Walter van Heuven
}
