% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{ngram_frequency_str}
\alias{ngram_frequency_str}
\title{ngram_frequency_str}
\usage{
ngram_frequency_str(
  the_str,
  ngram_table,
  type = "bigram",
  position_specific = TRUE,
  frequency = "token",
  func = "summed"
)
}
\arguments{
\item{the_str}{letter string}

\item{ngram_table}{table with ngram frequencies, use function `get_ngram_frequencies` for this}

\item{type}{monogram, bigram, or trigram}

\item{position_specific}{ngrams are position specific or not: TRUE or FALSE}

\item{frequency}{type or token}

\item{func}{summed}
}
\value{
summed ngram frequency
}
\description{
`ngram_frequency_str` calculates the ngram frequency of a single string
}
\author{
Walter van Heuven
}
