% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{ngrams}
\alias{ngrams}
\title{ngrams}
\usage{
ngrams(the_str, type = "bigram", frequency = 1)
}
\arguments{
\item{the_str}{letter string}

\item{type}{monogram, bigram, or trigram}

\item{frequency}{frequency of the letter string (word)}
}
\value{
list of ngram, position, type frequency, token frequency
}
\description{
`ngrams` returns letters, bigrams, or trigrams of a string with frequencies added
}
\author{
Walter van Heuven
}
