% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{sbf_rank}
\alias{sbf_rank}
\title{Summed Bigram Frequency (SBF) Rank}
\usage{
sbf_rank(the_str, bigram_table, top12 = FALSE, method = "Novick")
}
\arguments{
\item{bigram_table}{bigram table}

\item{top12}{TRUE or FALSE}

\item{method}{method to be used}

\item{the_Str}{the letter string}
}
\value{
list with SBF, SBF rank, and anagrams with SBF >= SBF the_str
}
\description{
Based on definition found in Novick & Sherman (2004)
}
