% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{str_bigrams}
\alias{str_bigrams}
\title{str_bigrams}
\usage{
str_bigrams(the_str, type = "adjacent", max_distance = -1)
}
\arguments{
\item{the_str}{a string}

\item{type}{'all', 'adjacent', 'open'}

\item{max_distance}{distance between the letters of the open bigrams. -1 no restriction for the number of open bigrams}
}
\value{
list of characters
}
\description{
`str_bigrams` returns the bigrams of a string
}
\details{
strngrams

Either all possible bigrams or only adjacent bigrams or open bigrams are returned
}
\examples{
\dontrun{
str_bigram("DREAM")
str_bigram("DREAM", "all")
}
}
\author{
Walter van Heuven, \email{walter.vaheuven@nottingham.ac.uk}
}
