% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{str_calc_ngrams}
\alias{str_calc_ngrams}
\title{str_calc_ngrams}
\usage{
str_calc_ngrams(
  word_list,
  freq_list,
  type = "bigram",
  position_specific = TRUE
)
}
\arguments{
\item{word_list}{list of words}

\item{freq_list}{list of frequencies for each word}

\item{type}{monogram, bigram, or trigram}

\item{position_specific}{ngram is position specific or not: TRUE or FALSE}
}
\value{
data.frame with a table of ngrams
}
\description{
`str_calc_ngrams` caclulates the ngrams of the strings in a list
}
\author{
Walter van Heuven
}
