% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{str_ngrams}
\alias{str_ngrams}
\title{str_ngrams}
\usage{
str_ngrams(the_str, type = "bigram", frequency = 1)
}
\arguments{
\item{the_str}{string of characters}

\item{type}{monogram, bigram, or trigram}

\item{frequency}{frequency of the the_str}
}
\value{
list of ngram, position, type frequency, token frequency
}
\description{
`str_ngrams` returns monograms, bigrams, or trigrams of a string with frequency information
}
\author{
Walter van Heuven, \email{walter.vaheuven@nottingham.ac.uk}
}
