% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{anagrams}
\alias{anagrams}
\title{Create all possible anagrams of the letter string}
\usage{
anagrams(the_str)
}
\arguments{
\item{the_str}{string of letters}
}
\value{
vector of all possible anagrams
}
\description{
Create all possible anagrams of the letter string
}
