% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{anagrams}
\alias{anagrams}
\title{anagrams of a string}
\usage{
anagrams(the_str, wordList = NULL, progressbar = TRUE)
}
\arguments{
\item{the_str}{string}

\item{wordList}{list of words}

\item{progressbar}{show progress bar TRUE or FALSE}
}
\value{
vector of all possible anagrams
}
\description{
`anagrams` returns all possible anagrams of the string
}
\details{
Note that this function uses a brute force method so use only short strings (<12)
}
\examples{
\dontrun{
# return all anagrams that are words
anagrams("dream", vwr::english.words)

# return all anagram letter strings, if none is found return NULL
anagrams("dream")
}
}
\author{
Walter van Heuven
}
