% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{anagrams}
\alias{anagrams}
\title{anagrams}
\usage{
anagrams(the_str)
}
\arguments{
\item{the_str}{string of letters}
}
\value{
vector of all possible anagrams
}
\description{
`anagrams` returns all possible anagrams of the letter string
}
\examples{
\dontrun{
anagrams("DREAM")
}
}
\author{
Walter van Heuven, \email{walter.vaheuven@nottingham.ac.uk}
}
