% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{bigrams}
\alias{bigrams}
\title{bigrams of a character string}
\usage{
bigrams(the_str, type = "adjacent", max_distance = -1)
}
\arguments{
\item{the_str}{character string}

\item{type}{'all', 'adjacent', 'open'}

\item{max_distance}{distance between the letters of the open bigrams. -1 no restriction for the number of open bigrams}
}
\value{
list of characters
}
\description{
`bigrams` return the bigrams of a letter string
}
\details{
strngrams

Written by Walter van heuven, https://waltervanheuven.net

Either all possible bigrams or only adjacent bigrams or open bigrams are returned
}
\examples{
\dontrun{
# only adjacent bigrams
bigram("dream")
# all possible open bigrams
bigram("dream", "all")
# max two letters between letters of the open bigram
bigram("dream", "open", 3)
}
}
\author{
Walter van Heuven
}
