% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{get_ngram_frequencies}
\alias{get_ngram_frequencies}
\title{get_ngram_frequencies}
\usage{
get_ngram_frequencies(
  word_list,
  freq_list,
  type = "bigram",
  position_specific = TRUE
)
}
\arguments{
\item{word_list}{list of words}

\item{freq_list}{list of frequencies for each word}

\item{type}{monogram, bigram (default), or trigram}

\item{position_specific}{ngram is position specific or not: TRUE (default) or FALSE}
}
\value{
data.frame with a table of ngrams
}
\description{
`get_ngram_frequencies` returns the ngrams of a lexicon
}
\author{
Walter van Heuven
}
