% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{gtzero}
\alias{gtzero}
\title{gtzero}
\usage{
gtzero(the_str, bigram_table)
}
\arguments{
\item{the_str}{the string}

\item{bigram_table}{bigram table}
}
\value{
list of bigrams
}
\description{
`gtzero` Returns the number of bigrams in the_str that exist in the bigram_table.
}
\details{
Based on definition in Mendelsohn (1976)
}
