% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{sbf_rank}
\alias{sbf_rank}
\title{sbf_rank}
\usage{
sbf_rank(the_str, bigram_table, top12 = FALSE, method = "Novick")
}
\arguments{
\item{the_str}{letter string}

\item{bigram_table}{bigram table}

\item{top12}{TRUE or FALSE}

\item{method}{method to be used, e.g. Novick (default)}
}
\value{
list with SBF, SBF rank, and anagrams with SBF >= SBF the_str
}
\description{
`sbf_rank` returns the summed Bigram Frequency (SBF) Rank
}
\details{
Based on definition found in Novick & Sherman (2004)
}
\references{
\url{https://link.springer.com/article/10.3758/BF03195587}
}
\author{
Walter van Heuven, \email{walter.vaheuven@nottingham.ac.uk}
}
