% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{str_ngram_frequency}
\alias{str_ngram_frequency}
\title{str_ngram_frequency}
\usage{
str_ngram_frequency(
  the_str,
  ngram_table,
  type = "bigram",
  position_specific = TRUE,
  frequency = "token",
  func = "summed"
)
}
\arguments{
\item{the_str}{the character string}

\item{ngram_table}{table with ngrams}

\item{position_specific}{TRUE or FALSE}

\item{frequency}{type or token}

\item{func}{summed}
}
\value{
summed ngram frequency
}
\description{
`str_ngram_frequency` caculates the ngram frequency of a string
}
