% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strngrams.R
\name{str_ngram_info}
\alias{str_ngram_info}
\title{str_ngram_info}
\usage{
str_ngram_info(
  word_list,
  ngram_table,
  type = "bigram",
  position_specific = TRUE,
  frequency = "token",
  func = "summed"
)
}
\arguments{
\item{word_list}{list of words}

\item{ngram_table}{ngram table}

\item{type}{monogram, bigram, or trigram}

\item{position_specific}{TRUE or FALSE}

\item{frequency}{type or token}

\item{func}{summed (default) or mean frequency}
}
\value{
vector of frequencies
}
\description{
`str_ngram_info` returns ngram info for each word (summed frequency) in a list
}
\author{
Walter van Heuven
}
