\name{strngrams}
\alias{strngrams}
\title{strngrams}
\usage{
Usage information.
}
\description{
Functions to extract ngrams (e.g., letters, bigrams, trigrams), create anagrams, and calculate summed/mean ngram type and token frequencies of letter strings.
}
\examples{
str_bigrams("DREAM")

str_bigrams("DREAM", "open")

str_ngram("DREAM")
}
